\docType{class}
\name{errorHandler-class}
\alias{errorHandler-class}
\title{CLASS errorHandler}
\description{
class to handle errors
}
\details{
brief desc here
}
\section{Methods}{
  \describe{ \item{\code{new(...)}:}{ Create a new
  \code{errorHandler} object. ... is used to define the
  appropriate slots.} }
}
\examples{
\dontrun{
}
}

