% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{latestChromeDriver}
\alias{latestChromeDriver}
\title{Get latest ChromeDriver version number for a specific Google Chrome /
Chromium release}
\usage{
latestChromeDriver(version)
}
\arguments{
\item{version}{A Google Chrome / Chromium version number consisting of major
-- and optionally minor, and build -- version number(s). A character scalar.}
}
\value{
A character scalar.
}
\description{
This function adheres to the version detection principles \href{https://chromedriver.chromium.org/downloads/version-selection}{specified by the ChromeDriver project}.
}
\examples{
latestChromeDriver("87")
latestChromeDriver("87.0.4280")
}
