\docType{data}
\name{selKeys}
\alias{selKeys}
\title{Selenium key mappings}
\format{A named list. The names are the descriptions of the keys. The values are the "UTF-8" character representations.}
\source{
  http://code.google.com/p/selenium/wiki/JsonWireProtocol#/session/:sessionId/element/:id/value
}
\description{
  This data set contains a list of selenium key mappings.
  The key mappings are outlined
  http://code.google.com/p/selenium/wiki/JsonWireProtocol#/session/:sessionId/element/:id/value.
}
\author{
  John Harrison, 2012-10-05
}
\keyword{datasets}

