% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{systemChromeVersion}
\alias{systemChromeVersion}
\title{Get installed stable Google Chrome version}
\usage{
systemChromeVersion()
}
\value{
A character scalar.
}
\description{
This function returns the first three elements of the (usually) four-element-long
version number of the stable Google Chrome binary detected on the system.
}
