% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeFirefoxProfile}
\alias{makeFirefoxProfile}
\title{Make Firefox profile.}
\usage{
makeFirefoxProfile(opts)
}
\arguments{
\item{opts}{option list of firefox}
}
\description{
\code{makeFirefoxProfile}
A utility function to make a firefox profile.
}
\note{
Windows doesn't come with command-line zip capability. Installing rtools
\url{http://cran.r-project.org/bin/windows/Rtools/index.html} is a straightforward way to gain
this capability.
}
\section{Detail}{
 A firefox profile directory is zipped and base64 encoded. It can then be passed
to the selenium server as a required capability with key firefox_profile
}
\examples{
\dontrun{
fprof <- makeFirefoxProfile(list(browser.download.dir = "D:/temp"))
remDr <- remoteDriver(extraCapabilities = fprof)
remDr$open()
}
}

