% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selKeys-data.R
\docType{data}
\name{selKeys}
\alias{selKeys}
\title{Selenium key mappings}
\format{A named list. The names are the descriptions of the keys. The
   values are the "UTF-8" character representations.}
\source{
https://github.com/SeleniumHQ/selenium/wiki/JsonWireProtocol#sessionsessionidelementidvalue
}
\usage{
selKeys
}
\description{
This data set contains a list of selenium key mappings.
selKeys is used when a sendKeys variable is needed.
sendKeys is defined as a list.
If an entry is needed from selKeys it is denoted by key.
}
\keyword{datasets}
