\name{startServer}
\alias{startServer}
\title{Start the standalone server.}
\usage{
startServer(dir = NULL)
}
\arguments{
  \item{dir}{A directory in which the binary is to be
  placed.}
}
\description{
\code{startServer} A utility function to start the
standalone server.
}
\section{Detail}{
  By default the binary is assumed to be in the RSelenium
  package /bin directory.
}
\examples{
\dontrun{
startServer()
}
}

